Updating a User's Username
Public void method that takes usernames and the change that the user would like to make has been added; it also checks (and prints to console) if the user is already taken (invalid username) to determine if it can make the change.

A setter for User usernames has also been included in User.java.