Faulty passcode invalidation corrected
Reordering of passcode validation conditions to properly indicate issues with changing a password was implemented. First checking to see if the old passcode field is correct, then if the new passcode is empty, next if the new and confirm passcodes match, and lastly, that the new passcode isn't the same as the old one.