Removed "Daily Workout Tips" + UI Update
Was discussed and found to be an unnecessary component for a program of such a nature; was more of a gimmick than anything.

Uniform look has been set up on dashboard and other screens (instructor registration screen, authentication code screen, admin registration screen, etc.).

Switching from 'Register' screen to 'Sign In' screen has been updated to wipe the password as well.