UI Overhaul + Minor Misc Fixes
Major changes to components, UI helper methods, and restructuring of JFrames to ensure UI consistency across different screens.

DashboardFrame now handles two different views (one for Users and Instructors, another for the Admin account).

Background was brought to the UserInfoFrame, such that the background is consistent. Resetting fields as tabs are traversed was also included for the UserInfoFrame (DetailsPanel and PassChangePanel).

Some images were also corrected.